Create a simple layout of images

Create a simple sequential layout of images

Usage

pixfigure(
  path,
  caption = NULL,
  link = TRUE,
  h = "auto",
  w = "100%",
  fit = "cover",
  position = "center",
  width = "100%",
  height = "100%",
  elementId = NULL
)

Arguments

path

A character vector of full paths to images.

caption

A character vector of captions for the images (Optional).

link

A logical or character vector. What happens when you click on the image? TRUE opens up the lightbox, FALSE to disable the lightbox. A character vector of custom URLs equal to length of path.

h

A character denoting height of the image in valid CSS units.

w

A character denoting width of the image in valid CSS units.

fit

String. Passed to object-fit CSS property.

position

String. Passed to object-position CSS property.

width

A character denoting width of the widget as a string in valid CSS units.

height

A character denoting height of the widget as a string in valid CSS units.

elementId

A character string denoting parent container ID.

Examples

library(pixture)
paths <- c(
  "https://images.pexels.com/photos/572897/pexels-photo-572897.jpeg",
  "https://images.pexels.com/photos/7604425/pexels-photo-7604425.jpeg",
  "https://images.pexels.com/photos/4666748/pexels-photo-4666748.jpeg"
)
pixfigure(paths)

# local example
if (FALSE) { # \dontrun{
library(pixture)
paths <- list.files(path=system.file("extdata/images",package="pixture"),full.names=TRUE)
pixfigure(paths)
} # }