Shiny bindings for pixgallery

Output and render functions for using pixgallery within Shiny applications and interactive documents.

Usage

pixgalleryOutput(outputId, width = "100%", height = "100%")

renderPixgallery(expr, env = parent.frame(), quoted = FALSE)

Arguments

outputId

output variable to read from

width, height

Must be a valid CSS unit (like '100%', '400px', 'auto') or a number, which will be coerced to a string and have 'px' appended.

expr

An expression that generates a pixgallery

env

The environment in which to evaluate expr.

quoted

Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.