Create an image gallery.

Create an image gallery.

pixture(
  path,
  caption = NULL,
  preset = "grid",
  options = list(),
  width = "100%",
  height = NULL,
  elementId = NULL
)

Arguments

path
A character vector of paths to images.

caption
A character vector of captions for the images (Optional).

preset
A character denoting predefined layout style. Defaults to "grid".

options
A named list of freewall options, See http://kombai.github.io/freewall/#options.

width
A numeric denoting width of the widget.

height
A numeric in pixel denoting height of the widget.

elementId
A character denoting parent container ID.